+++ relay-server/src/processing/logs/store.rs
@@ -2,15 +2,15 @@ use std::collections::HashMap;
 
 use chrono::{DateTime, Utc};
 use relay_event_schema::protocol::{Attributes, OurLog, OurLogLevel, SpanId};
-use relay_protocol::{Annotated, IntoValue, Value};
+use relay_protocol::Annotated;
 use relay_quotas::Scoping;
 use sentry_protos::snuba::v1::{AnyValue, TraceItem, TraceItemType, any_value};
 use uuid::Uuid;
 
 use crate::envelope::WithHeader;
 use crate::processing::logs::{Error, Result};
-use crate::processing::utils::store::{AttributeMeta, extract_meta_attributes, proto_timestamp};
-use crate::processing::{Counted, Retention};
+use crate::processing::utils::store::{extract_meta_attributes, proto_timestamp};
+use crate::processing::{self, Counted, Retention};
 use crate::services::outcome::DiscardReason;
 use crate::services::store::StoreTraceItem;
 
@@ -114,42 +114,7 @@ fn attributes(
     // +N, one for each field attribute added and some extra for potential meta.
     result.reserve(attributes.0.len() + 5 + 3);
 
-    for (name, attribute) in attributes {
-        let meta = AttributeMeta {
-            meta: IntoValue::extract_meta_tree(&attribute),
-        };
-        if let Some(meta) = meta.to_any_value() {
-            result.insert(format!("sentry._meta.fields.attributes.{name}"), meta);
-        }
-
-        let value = attribute
-            .into_value()
-            .and_then(|v| v.value.value.into_value());
-
-        let Some(value) = value else {
-            // Meta has already been handled, no value -> skip.
-            // There are also no current plans to handle `null` in EAP.
-            continue;
-        };
-
-        let Some(value) = (match value {
-            Value::Bool(v) => Some(any_value::Value::BoolValue(v)),
-            Value::I64(v) => Some(any_value::Value::IntValue(v)),
-            Value::U64(v) => i64::try_from(v).ok().map(any_value::Value::IntValue),
-            Value::F64(v) => Some(any_value::Value::DoubleValue(v)),
-            Value::String(v) => Some(any_value::Value::StringValue(v)),
-            // These cases do not happen, as they are not valid attributes
-            // and they should have been filtered out before already.
-            Value::Array(_) | Value::Object(_) => {
-                debug_assert!(false, "unsupported log value");
-                None
-            }
-        }) else {
-            continue;
-        };
-
-        result.insert(name, AnyValue { value: Some(value) });
-    }
+    processing::utils::store::convert_attributes_into(&mut result, attributes);
 
     let FieldAttributes {
         level,